requirements.txt
accelerate>=1.0.1
datasets>=3.0.0
faiss-cpu>=1.8.0
langgraph>=0.2.30
matplotlib>=3.8.0
mongomock>=4.1.2
networkx>=3.2
numpy>=1.26.0
pandas>=2.2.0
peft>=0.19.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
pymongo>=4.8.0
pypdf>=5.0.0
python-docx>=1.1.2
python-dotenv>=1.0.1
pytest>=8.3.0
scikit-learn>=1.5.0
seaborn>=0.13.2
streamlit>=1.55.0
torch>=2.4.0
transformers>=4.45.0
watchdog>=5.0.0
sentence-transformers>=3.0.1
